Implemented listing of saved shops & changed column type in first migration
NOTE: `php artisan migrate:refresh` or `php artisan migrate:fresh` needed to be run.

+++ database/migrations/2018_07_19_194706_create_shops_table.php
@@ -18,9 +18,9 @@ class CreateShopsTable extends Migration
 
             $table->string('name');
             $table->string('address');
-            $table->integer('distance');
-            $table->float('latitude');
-            $table->float('longitude');
+            $table->float('distance');
+            $table->decimal('latitude', 12, 8);
+            $table->decimal('longitude', 12, 8);
 
             $table->timestamps();
         });
